[ndproj] Fixing bug when deleting an empty redis channel

+++ ndproj/redisprojdb.py
@@ -28,8 +28,15 @@ class RedisProjectDB:
     """Create the database connection"""
     self.pr = NDProject(project_name)
     # Connect to the redis cluster
-    self.client = redis.StrictRedis(host=self.pr.getDBHost(), port=6379, db=0)
-    self.pipe = self.client.pipeline(transaction=False)
+    try:
+      self.client = redis.StrictRedis(host=self.pr.getDBHost(), port=6379, db=0)
+      self.pipe = self.client.pipeline(transaction=False)
+    except redis.ConnectionError as e:
+      logger.error("Cannot connect to Redis server. {}".format(e))
+      raise NDWSError("Cannot connect to Redis server. {}".format(e))
+    except Exception as e:
+      logger.error("Unknown error while connecting to Redis. {}".format(e))
+      raise NDWSError("Unknown error while connecting to Redis. {}".format(e))
 
 
   def close (self):
@@ -53,9 +60,14 @@ class RedisProjectDB:
     # KL TODO Is this redundant?
     # project pattern to fetch all the keys with project_name
     project_pattern = "{}_*".format(self.pr.getProjectName())
-    project_keys = self.client.keys(project_pattern)
-    # delete all the keys with the pattern
-    self.client.delete(*project_keys)
+    try:
+      project_keys = self.client.keys(project_pattern)
+      # delete all the keys with the pattern
+      if project_keys:
+        self.client.delete(*project_keys)
+    except Exception as e:
+      logger.error("Error in deleting Redis project {}. {}".format(self.pr.getProjectName(), e)
+      raise NDWSError("Error in deleting Redis project {}. {}".format(self.pr.getProjectName(), e)
 
 
   def deleteNDChannel(self, channel_name):
@@ -64,6 +76,11 @@ class RedisProjectDB:
     # KL TODO Maybe do this as a transaction?
     # channel pattern to fetch all the keys with project_name_channel_name
     channel_pattern = "{}_{}_*".format(self.pr.getProjectName(), channel_name)
-    channel_keys = self.client.keys(channel_pattern)
-    # delete all the keys with the pattern
-    self.client.delete(*channel_keys)
+    try:
+      channel_keys = self.client.keys(channel_pattern)
+      # delete all the keys with the pattern
+      if channel_keys:
+        self.client.delete(*channel_keys)
+    except Exception as e:
+      logger.error("Error in deleting channel {}. {}".format(channel_name, e)
+      raise NDWSError("Error in deleting channel {}. {}".format(channel_name, e)
